fix student input simulations
